requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
langchain>=0.0.350
langchain-community>=0.0.10
langchain-huggingface>=0.0.5
faiss-cpu>=1.7.4
mistralai>=0.0.11
numpy>=1.26.2
scikit-learn>=1.3.2
sentence-transformers>=2.2.2 